+++ UserSearch/UserSearch.js
@@ -2,12 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Button from '@folio/stripes-components/lib/Button';
 import Icon from '@folio/stripes-components/lib/Icon';
+import className from 'classnames';
 
 import css from './UserSearch.css';
 import UserSearchModal from './UserSearchModal';
 
 export default class UserSearch extends React.Component {
-
   constructor(props) {
     super(props);
 
@@ -17,6 +17,15 @@ export default class UserSearch extends React.Component {
 
     this.openModal = this.openModal.bind(this);
     this.closeModal = this.closeModal.bind(this);
+  }
+
+  getStyle() {
+    const { marginBottom0, marginTop0 } = this.props;
+    return className(
+      css.searchControl,
+      { [css.marginBottom0]: marginBottom0 },
+      { [css.marginTop0]: marginTop0 },
+    );
   }
 
   openModal() {
@@ -33,12 +42,14 @@ export default class UserSearch extends React.Component {
 
   render() {
     return (
-      <div className={css.searchControl}>
+      <div className={this.getStyle()}>
         <Button
+          id="clickable-plugin-find-user"
           key="searchButton"
           buttonStyle={this.props.searchButtonStyle}
           onClick={this.openModal}
-          title="Find User" tabIndex="-1"
+          title="Find User"
+          tabIndex="-1"
         >
           {this.props.searchLabel ? this.props.searchLabel : <Icon icon="search" color="#fff" />}
         </Button>
@@ -50,7 +61,6 @@ export default class UserSearch extends React.Component {
       </div>
     );
   }
-
 }
 
 UserSearch.defaultProps = {
@@ -60,4 +70,6 @@ UserSearch.defaultProps = {
 UserSearch.propTypes = {
   searchLabel: PropTypes.string,
   searchButtonStyle: PropTypes.string,
+  marginBottom0: PropTypes.bool,
+  marginTop0: PropTypes.bool,
 };