+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.2.0] (IN PROGRESS)
 
 * Import PropTypes from prop-types, not React. Refs STRIPES-427. 
+* Lint cleanup. Fixes UIPFU-3.
 
 ## [1.1.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.1.0) (2017-09-06)