+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Import PropTypes from prop-types, not React. Refs STRIPES-427.
 * Lint cleanup. Fixes UIPFU-3.
 * Add `dataKey` to connected Users component. Fixes UIPFU-4.
+* Use more-current stripes-components. Refs STRIPES-495.
 
 ## [1.1.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.1.0) (2017-09-06)