+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Use more-current stripes-components. Refs STRIPES-495.
 * Use more-current stripes-connect. Refs STRIPES-501.
 * Turn SearchAndSort's show-single-result feature off. Refs STSMACOM-52. Fixes UIREQ-60, UICHKOUT-54, UIU-373.
+* Pass `browseOnly` to `<Users>` to suppress user creation and editing. Refs UIPFU-6. Available from v1.1.1.
 
 ## [1.1.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.1.0) (2017-09-06)
 
@@ -17,5 +18,3 @@
 ## [1.0.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.0.0) (2017-07-03)
 
 * First release.
-
-