+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Use more-current stripes-connect. Refs STRIPES-501.
 * Turn SearchAndSort's show-single-result feature off. Refs STSMACOM-52. Fixes UIREQ-60, UICHKOUT-54, UIU-373.
 * Pass `browseOnly` to `<Users>` to suppress user creation and editing. Refs UIPFU-6. Available from v1.1.1.
+* Set element ID on button for regression testing. UICHKOUT-414.
 
 ## [1.1.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.1.0) (2017-09-06)