+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * Import PropTypes from prop-types, not React. Refs STRIPES-427.
 * Lint cleanup. Fixes UIPFU-3.
 * Add `dataKey` to connected Users component. Fixes UIPFU-4.
+* Use more-current stripes-components. Refs STRIPES-495.
+* Use more-current stripes-connect. Refs STRIPES-501.
+* Turn SearchAndSort's show-single-result feature off. Refs STSMACOM-52. Fixes UIREQ-60, UICHKOUT-54, UIU-373.
+* Pass `browseOnly` to `<Users>` to suppress user creation and editing. Refs UIPFU-6. Available from v1.1.1.
+* Set element ID on button for regression testing. UICHKOUT-414.
 
 ## [1.1.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.1.0) (2017-09-06)
 
@@ -14,5 +19,3 @@
 ## [1.0.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.0.0) (2017-07-03)
 
 * First release.
-
-