+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-plugin-find-user
 
-## [1.2.0] (IN PROGRESS)
+## [1.2.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.2.0) (2018-09-04)
+[Full Changelog](https://github.com/folio-org/ui-plugin-find-user/compare/v1.1.0...v1.2.0)
 
 * Import PropTypes from prop-types, not React. Refs STRIPES-427.
 * Lint cleanup. Fixes UIPFU-3.
@@ -10,8 +11,11 @@
 * Turn SearchAndSort's show-single-result feature off. Refs STSMACOM-52. Fixes UIREQ-60, UICHKOUT-54, UIU-373.
 * Pass `browseOnly` to `<Users>` to suppress user creation and editing. Refs UIPFU-6. Available from v1.1.1.
 * Set element ID on button for regression testing. UICHKOUT-414.
+* Import Users without routing. Fixes UIPFU-7.
+* Make styles more flexible. Fixes UIPFU-5.
 
 ## [1.1.0](https://github.com/folio-org/ui-plugin-find-user/tree/v1.1.0) (2017-09-06)
+[Full Changelog](https://github.com/folio-org/ui-plugin-find-user/compare/v1.0.0...v1.1.0)
 
 * Update stripes-components to 1.7.0.
 * Add ability to customize search button. Fixes UIPFU-2.